requirements.txt
discord.py==1.4.1
python-dotenv==0.14.0
flask==1.0.2
PyDrive==1.3.1